RAG Model/requirements.txt
# # # Core LangChain and Google Gemini
# # langchain==0.3.7
# # langchain-community==0.3.7
# # langchain-google-genai==2.1.12
# # google-generativeai==0.8.3

# # # Vector Database and Embeddings
# # faiss-cpu==1.12.0
# # sentence-transformers==5.1.1

# # # Keyword Search
# # rank-bm25==0.2.2

# # # Document Processing
# # pypdf==4.3.1
# # python-docx==1.1.2
# # python-dotenv==1.0.1

# # # Text Processing and Utilities - Python 3.13 Compatible
# # tiktoken==0.8.0
# # numpy>=2.1.0  # Python 3.13 compatible version
# # pandas>=2.2.0

# # # Memory and Caching
# # diskcache==5.6.3

# # # Web Interface (Optional)
# # streamlit==1.39.0
# # streamlit-chat==0.1.1

# # # Development and Testing
# # pytest==8.3.3
# # python-decouple==3.8

# # Core LangChain and Google Gemini - Updated versions
# langchain>=0.3.19
# langchain-community>=0.3.19
# langchain-core>=0.3.19
# langchain-google-genai>=2.1.12
# google-generativeai>=0.8.3

# # Vector Database and Embeddings
# faiss-cpu>=1.12.0
# sentence-transformers>=5.1.1

# # Keyword Search
# rank-bm25>=0.2.2

# # Document Processing
# pypdf>=4.3.1
# python-docx>=1.1.2
# python-dotenv>=1.0.1

# # Text Processing and Utilities - NumPy 2.x compatible
# tiktoken>=0.8.0
# numpy>=2.1.0
# pandas>=2.2.0

# # Memory and Caching
# diskcache>=5.6.3

# # Web Interface (Optional)
# streamlit>=1.39.0
# streamlit-chat>=0.1.1

# # Development and Testing
# pytest>=8.3.3
# python-decouple>=3.8

# Core LangChain components
langchain>=0.3.19
langchain-community>=0.3.19
langchain-core>=0.3.19

# Google Gemini Integration - Compatible versions
langchain-google-genai>=2.1.12
# Remove google-generativeai - it's automatically included as dependency

# Vector Database and Embeddings
faiss-cpu>=1.12.0
sentence-transformers>=5.1.1

# Keyword Search
rank-bm25>=0.2.2

# Document Processing
pypdf>=4.3.1
python-docx>=1.1.2
python-dotenv>=1.0.1

# Text Processing and Utilities
tiktoken>=0.8.0
numpy>=2.1.0
pandas>=2.2.0

# Memory and Caching
diskcache>=5.6.3

# Web Interface (Optional)
streamlit>=1.39.0
streamlit-chat>=0.1.1

# Development and Testing
pytest>=8.3.3
python-decouple>=3.8
google_genai
# Google Gemini API
google-generativeai>=0.8.3

# Data processing
pandas>=2.0.0

# PDF generation
reportlab>=4.0.0

# Optional: For additional data manipulation
numpy>=1.24.0